improved distribution calculation
traffic source
--------------

the referer ('traffic source') is calculated as follows: for each unique search the earliest hit with an external referer for the current month. count the occurance of domains in those referers.

so each unique can only contribute "+1" to one and only one referer. even if the unique came 10x with the same referer or 10x with 10 different referers (in the later case, the first referer is counted).

the above calculation could be done in SQL but this was terrible slow on my machine even with indizes on `unique` and `day`:

    select hit1.userAgent, hit1.referer, hit1.page from Hit as hit1
    left join Hit as hit2 
    on (hit2.month = :month and hit2.site = :site and
        hit1.unique = hit2.unique and hit1.day < hit2.day
    )
    where hit2.unique is null
    

similarly, the distinct userAgents could be queried via SQL but i would still have to do normalization in JS so this doesn't seem worth the coding effort.